feat: Refactor: Move theme related files to designsystem package
This commit refactors the project structure by moving theme-related files into a `designsystem` package for better organization and separation of concerns.

- **Package Reorganization**:
    - Moved `Color.kt`, `Theme.kt`, and `Type.kt` from the `ui.theme` package to the `designsystem.theme` package.
- **Import Updates**:
    - Updated imports in `ErrorStateView.kt`, `Header.kt`, `MainActivity.kt`, and `SectionContainer.kt` to reflect the new package structure for theme-related classes.
- **Deleted Files**:
    - Removed `ImagePager.kt` and `ImagePagerContent.kt`.
