+++ app/src/main/java/com/example/stylefeed/designsystem/components/ErrorStateView.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.stylefeed.ui.theme.FashionContentTheme
+import com.example.stylefeed.designsystem.theme.FashionContentTheme
 
 @Composable
 fun ErrorStateView(
