+++ app/src/main/java/com/example/stylefeed/designsystem/components/header/Header.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil.compose.rememberAsyncImagePainter
-import com.example.stylefeed.ui.theme.FashionContentTheme
+import com.example.stylefeed.designsystem.theme.FashionContentTheme
 
 
 @Composable
