+++ app/src/main/java/com/example/stylefeed/designsystem/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.stylefeed.ui.theme
+package com.example.stylefeed.designsystem.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
